Update User.php
Fix info() to get currentOnlineId which will hold information about the new onlineId if the user have changed it. It will not be returned at all if the user hasn't changed onlineId. Noticed that my.playstation.com also asks about displayableOldOnlineId, which I also included in this change, but I haven't found anything that will return this value.