Get token id from users search result This token can be "token_invalidated:deleted", which I assume means that the player have been deleted by Sony.
Making further calls on the accountId will just net an error:
"code": 2281485,
"message": "Resource not found (account='9068770825948628755')"

The user I encountered this upon was "ChaoticChaos", with id 9068770825948628755.

Also changed how onlineId is fetched because of above mentioned error.

And found out that max pageSize for search result is 50, so that's the default now.